cmd/snap-bootstrap/initramfs-mounts: call systemd-mount instead of the-tool
When we reach the end of the initramfs, we used to do an isolate of the various
initrd systemd units, which would delete all units that were not direct
dependencies of the initrd units, and this would also in effect unmount some of
the things that we requested be mounted from snap-bootstrap leading to other
problems. The intermediate solution to this was to switch to using start instead
of isolate, this allowed the mount units to persist, but is not desirable as
services and other things running in the initramfs do not release freed memory,
as well as leaking various state to the running system which we probably do not
want leaked.

Instead, what we do now is we will mount things manually from snap-bootstrap,
and create override files for the initrd systemd units to depend on the mounts
we create, thus preventing them from being unmounted and disappearing after we
leave the initramfs. Doing this dynamically ensures that as we need to mount
additional things from the initramfs, we do not need to make changes to the
systemd static file layout in the ubuntu-core-initramfs package and instead we
just start mounting it here in the code and it works as expected.

Finally, since we want accurate testing of this change before removing the start
instead of isolate workaround, we need to patch the initramfs when we build it
for tests with this version of snap-bootstrap to operate the way it is supposed
to. This patching should go away when upstream ubuntu-core-initramfs is fixed.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>