snapshotstate: rework snapshotstate.Export() to avoid the need for locks
This commit changes the export so that it collects open file
descriptors of all files that are part of the snapshot id. With
that the files can be served even if the snapshot is deleted
while the data is served because we have an open fd on each
file we need to server. This avoids the need to lock the export.