sandbox/cgroup: test delayed scope activation
In most recent versions of systemd, as used in the rolling-release
distributions such as openSUSE tumbleweed, we have observed that
StartTransientUnit performs the cgroup manipulation after a small
delay, long after returning from the D-Bus call.

This was handled in a bit of work-around logic but was not tested. This
patch corrects that with a mock time-dependent test.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>