many/tests/preseed: reset the preseeded images before preseeding them
The cloud images we consume for these tests are starting to be preseeded
themselves, so for now we reset any preseeding that might exist in the images
before doing our tests. Resetting the preseed is safe to do even if the image
wasn't preseeded, so this is always safe to do.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>