o/snapshotstate/backend/helpers.go: adjust UnknownUser{Id,}Error handling
As with osutil/user.FindUid(), we suffer from an upstream problem where error
handling surrounding positively identifying if a given user/group/id does not
exist is impossible. As such, treat all non-nil errors as the unknown case, in
this case what happens is that when creating snapshots, we may silently miss
some users if there are issues looking up their name/id, but that is better
behavior than today where we would not create any snapshot if there was an
errant /home/not-a-user/ directory and that not-a-user returned a non-0 errno
from getpwnam_r.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>