many: rename osutil.RealUser to osutil.UserMaybeSudoUser, add more comments
The name of this function was confusing, as it is not the real user in the sense
of real vs effective uid, but rather either the normal current user, or if we
are root via sudo method, then the user calling sudo.

Additionally, add a comment about the ineffectiveness of checking for
user.UnknownUserError here.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>