interfaces/system_key.go: add UnmarshalJSONSystemKey
This allows other packages to import a JSON system-key representation without
needing to export details of the system-key format. This will be used for
example from the `snap debug seeding` command, as the client will get
system-keys over JSON that they want to compare, without knowing the internal
details of the system-key.

Also add a test that writing the key to disk with WriteSystemKey gives a
matching system-key as the UnmarshalJSONSystemKey method when applied to
matching JSON.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>