osutil/disks: use xerrors to indicate a fs label wasn't found
It will be necessary in the secboot package to positively indicate that a FS
label was not found and that we didn't have some other error, as the first step
to unlocking is to check if there is an encrypted device. We currently identify
that by using /dev/disk/by-label/ubuntu-data-enc, but this is insecure and buggy
as it means if the boot disk was not encrypted (i.e. on a device with a TPM but
without FDE enabled/used), an attacker can attach a USB drive with a FS label of
ubuntu-data-enc (as opposed to the real boot disk with ubuntu-data label on it),
and now the system will fail to boot because the ubuntu-data-enc USB drive FS
label will not be unlockable. We will avoid this bug by checking if the disk
that we booted the kernel from (and mounted the first partition using) has a
partition with a FS label on it called ubuntu-data-enc, rather than any
partition with a FS label of ubuntu-data-enc.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>