interfaces/system-key: in WriteSystemKey during tests, don't call ParserFeatures
During tests, this would end up getting the parser features from the host rather
than from the mocked key, so on systems like LP buildd hosts with old kernels we
would end up modifying the mocked system key, which is wrong.

Also change the test that was failing on LP to use a system key which will be
different from the one on most modern Ubuntu systems to trigger the test failure
so that the change is proven effective.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>