corecfg: add "system.timezone" setting to the system settings
This commit adds support for setting the timezone via the snap
configuration system. This allows image customizations in UC20
that used to be done via direct manipulation of the "writable"
partition in UC16/UC18.

The image customization part will be fine as is but for the runtime
setting we also need to fix LP: #1650688