tests: detect unexpected xenial kernels
We are sometimes seeing unexpected cgroup configuration on some systems.
We suspect that there may be a bug in GCE image selection that
infrequently gives us a different "ubuntu" xenial image.

The invariant checker now understands spread systems and contains an
allow-list for xenial only. The scheme can be expanded or abandoned,
depending on the outcome.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>