tests: reboot to restart logind only when necessary
This is an improvement over the previous patch. This time we only reboot
if we suspect systemd-logind needs to be restarted. This happens across
the 245-246 upgrade or if we had to work around a bug present in older
versions.

This means we may reboot less often and thus avoid rebooting if
unnecessary, as it seems to be causing reboot loops.

This and the previous patch will fail on ubuntu core 16, as there we
need to apply a mount change to be able to work around a bug present in
older systemd, but the mount change is lost on reboot. I have no idea
how to fix that yet.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>