tests: apply logind workaround only if linger is broken
logind has a bug where linger doesn't work because systemd-logind
configuration file is missing StateDirectory entry. This was fixed
in later versions so this code should not run all the time.

Originally, in tests.session, this logic was triggered only when
loginctl enable-linger call actually fails. Replicate that to
avoid rebooting the system when not stricly necessary.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>