sandbox/cgroup: remove temporary workaround for multiple cgroup writers
This patch removes code added in ee673c8a06c7f57e3a030bcf721c6d340e987303
where - because at the time we had multiple cgroup writers - snap-confine
and systemd on request of snap-run. With this patch, snap run no longer
waits for systemd to complete the transition to the new scope-named
cgroup hierarchy. When there is only one writer we have no chance of
corrupting what systemd is doing so this should be safe.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>