tests: add support for linger on core16
This patch modifies the core snap so that session lingering works on
core16, just like it does on all the other versions of core now.

There are two parts to the patch:
 - support for dynamic fixing of linger is removed
 - core is modified to allow writable /var/lib/systemd/linger

Dynamic fixing of linger support is not fully supported as restarting
systemd-logind.service loses some of the state, specifically the set of
sessions that were active.

This patch can be further simplified with modified core snap, so that
additional workarounds are not required.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>