gadget,kernel: add new "kernel" and "gadget/edition" packages
This commit moves the kernel.yaml parsing into the kernel package
and extracts the gadget:editionNumber" helper into it's own
gadget/editon packages so that it can be shared between the
kernel and the gadget packages. Because meta/kernel.yaml is
optional it is not an error if it is missing.

Extra tests are also added.