tests: unmount FUSE file-systems from XDG runtime dir
When a FUSE file system is mounted, killing the corresponding process
does not unmount it. Instead the filesystem stays around in a broken
state until it is unmounted explicitly.

When tests.session restores a non-root service it stops the whole user
slice, which effectively terminates all processes belonging to the test
user. This can cause issues with the XDG document portal, if shutdown is
performed on a busy machine and kill-timeout kicks in.

To avoid it, explicitly unmount gvfs and XDG document portal before
stopping the slice.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>