logger: add support for setting snapd.debug=1 on kernel cmdline
While debugging a failure during UC20 seeding I noticed it is
hard to get debug data from snapd. I booted the system with:
```
dangerous systemd.debug-shell=1
```
but the
```
journalctl -u snapd
```
output did not contain that much information because debug
was not enabled. So this commit allows to enable debug via
the kernel commandline to allow easier seeding debugging.