cmd/snap-bootstrap/initramfs-mounts: call systemd-mount instead of the-tool (#9010)
* cmd/snap-bootstrap/initramfs-mounts: call systemd-mount instead of the-tool

When we reach the end of the initramfs, we used to do an isolate of the various
initrd systemd units, which would delete all units that were not direct
dependencies of the initrd units, and this would also in effect unmount some of
the things that we requested be mounted from snap-bootstrap leading to other
problems. The intermediate solution to this was to switch to using start instead
of isolate, this allowed the mount units to persist, but is not desirable as
services and other things running in the initramfs do not release freed memory,
as well as leaking various state to the running system which we probably do not
want leaked.

Instead, what we do now is we will mount things manually from snap-bootstrap,
and create override files for the initrd systemd units to depend on the mounts
we create, thus preventing them from being unmounted and disappearing after we
leave the initramfs. Doing this dynamically ensures that as we need to mount
additional things from the initramfs, we do not need to make changes to the
systemd static file layout in the ubuntu-core-initramfs package and instead we
just start mounting it here in the code and it works as expected.

Finally, since we want accurate testing of this change before removing the start
instead of isolate workaround, we need to patch the initramfs when we build it
for tests with this version of snap-bootstrap to operate the way it is supposed
to. This patching should go away when upstream ubuntu-core-initramfs is fixed.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* cmd/snap-bootstrap/initramfs-mounts: adjust comments about blocking

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* cmd/snap-bootstrap/initramfs-mounts: rm stdout var

It's not used anymore as we don't output anything specifically to stdout and
thus we don't need to mock anything here.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* cmd/snap-bootstrap/initramfs-mounts: also add override confs for local-fs.target

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* cmd/snap-bootstrap/initramfs-mounts: mock time.Now, IsMounted, fix timeout error

We should have been returning an error when we timeout waiting for the mount to
appear.

Additionally, we will need to mock time.Now() to write proper tests for this,
and we also need to go back to mounting osutil.IsMounted for tests that use the
actual doSystemdMount codepath.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* cmd/snap-bootstrap/initramfs-mounts: put opts after args, prefix with root-dir

We will need to mock this so we need to use dirs.GlobalRootDir to test this
properly.

Also it will be easier to test things if the what and where args always come
first.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* tests/cmd/snap-bootstrap/initramfs-mounts: stop mocking Stdout

We don't use Stdout for anything expected anymore, so we can stop this.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* tests/cmd/snap-bootstrap/initramfs-mounts: test real doSystemdMount codepath

This requires extensive mocking.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* cmd/snap-bootstrap/initramfs-mounts: adjust help message

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* cmd/snap-bootstrap/initramfs-mounts: don't fsck ubuntu-seed in run mode

This will lead to unnecessary reading from ubuntu-seed, which we want to keep as
pristine as possible, so don't fsck it every time the system is booted in run
mode. Note we still fsck during install and recover mode because we want it to
be consistent during those modes.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* cmd/s-b/initramfs-mounts: rename helper to mountPartitionMatchingKernelDisk

Also delete old TODOs.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* cmd/snap-bootstrap/initramfs-mounts: adjust step numbering

We no longer check if things are mounted first before mounting them, we just
assume they are not mounted and attempt to mount them.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* cmd/snap-bootstrap/initramfs-mounts: use --fsck=no for !opts.NeedsFsck

The default for --fsck is "on", so if NeedsFsck is false, then we should
explicitly turn it off with `--fsck=no`.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* snap-bootstrap: cosmetic "err := ...\nif err {" change

This commit change most of the new code that has the error style
```
err := ...
if err != nil {
```
to the more compact (and more common in our codebase) single line
style:
```
if err := ...; err != nil {
```

Co-authored-by: Michael Vogt <mvo@ubuntu.com>