configcore: use "virtual" configuration for system.timezone
This commit starts using the configuration hijacking code for
the "system.timezone" configuration. This configuration item
becomes "virtual" now. I.e. any Get() call will trigger snapd
to read the actual timezone of the system. And Set() will set
it to the system but not store the result in the state.