cmd/snap-confine: harden sc_dprintf
Security review by Jamie indicated that we should harden sc_dprintf and
make sure it is well guarded against short writes. I was wondering how
to do that correctly, without relying on the fragile length of the
format string, which can be longer than a valid formatted output.

This patch contains an idea based on vsnprintf's ability to return the
number of bytes necessary to format a string without performing the
actual work and allocating a buffer.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>