opengl: also allow access to /dev/dri/card1 and higher (LP: #1890783)
Allow access to all cards since a) this is common on hybrid systems, b)
ARM devices commonly have two devices (such as on the Raspberry Pi 4,
one for KMS and another that does not) and c) there is nothing saying
that /dev/dri/card0 is the default card or the application is currently
using. Furthermore, we already allow access to /dev/dri/renderD* which
correspond to the different cards.

In the future we may want to breakout the /dev/dri/renderD* into a less
privileged interface to support simple opengl clients with modern
kernels and libraries (see LP: #1890783 for more discussion).

References:
https://bugs.launchpad.net/snapd/+bug/1890783