tests: fix lxd test on 16.04
The "systemctl daemon-reload" command fails inside a nested
lxd container. We did not notice this for a long time because
the "prep-snapd-in-lxd.sh" script was runnin without "set -e"
because of an earlier mistake.

The daemon-reload command fails with:
```
Failed to execute operation: Connection timed out
```
but the failure seems to be have no conequences so let's continue
for now until this is fully understood.