interfaces/{docker,kubernetes}-support: support systemd cgroup driver
containerd and kubelet have the ability to move a container process into
a systemd-managed cgroup. This is now the recommend configuration since
it provides a single cgroup manager (systemd) in an effort to achieve
consistent views of resources.

microk8s currently perform runtime detection which will configure the
systemd cgroup driver on modern systems (eg Ubuntu 19.10 and later).

References:
- https://kubernetes.io/docs/setup/production-environment/container-runtimes/
- https://github.com/ubuntu/microk8s/issues/519
- https://github.com/ubuntu/microk8s/pull/520