mkversion.sh: do not use git describe --dirty
The most recent snapd snap build generates a version number:
2.46+git2.46.2.46

This indicates that the change in c8da36b that added --dirty
generated a "-dirty" version number and hence the check if
the version is a release failed.

It turns out that the hash of the vendor.json changed and
this caused the "dirty" flag in git. Given that it's time
consuming to do a point release because of issues like this
I would like to avoid using "--dirty" to deal with flukes
like this. Alternatively we need to include the "dirty"
flag in the auto-generated git versions too and watch out
for them so that future releases are not affected by this.

Something like:
```
@@ -82,9 +82,10 @@ if [ -z "$version_from_user" ] && [ "$version_from_git" != "" ] && [ "$version_f
         echo "Cannot generate version, there is a version from git and the changelog has a git version"
         exit 1
     else
-        revno=$(git describe --always --abbrev=7|cut -d- -f2)
-        commit=$(git describe --always --abbrev=7|cut -d- -f3)
-        v="${version_from_changelog}+git${revno}.${commit}"
+        revno=$(git describe --dirty --always --abbrev=7|cut -d- -f2)
+        commit=$(git describe --dirty  --always --abbrev=7|cut -d- -f3)
+        dirty=".$(git describe --dirty  --always --abbrev=7|cut -d- -f4)"
+        v="${version_from_changelog}+git${revno}.${commit}${dirty}"
         o="changelog+git"
     fi
 fi
```