mkversion.sh: simple hack to include dirty in version if the tree is dirty
We want dirty to show up in the version number if the tree is in fact dirty when
we build. Currently, if the version is dirty on a tag, we would get something
like 2.46-dirty as version_from_git, which will not be the same as 2.46 from the
changelog, and then we produce a silly version number like:

2.46+git2.46.2.46

but with this we now produce:

2.46-dirty

And for commits that are not directly on a tag, and are dirty, we get

2.46+git83.g1671726-dirty

which also produces the desired effect of including dirty in the version.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>