kernel: remove "edition" from kernel.yaml and add "update"
This commit changes the supported yaml for the kernel snap. For
the raspberry pi DTB use case the kernel assets are tightly
coupled with the kernel. So the edition would have to be bumped
everytime the kernel is build. So the edition does not make much
sense in this context. Hence a new "update" field that is boolean
for now but we may expand it later into a map. This map would
allow to specify what content items should get updated and which
should not get updated.