o/devicestate/devicemgr.go: disable NoCloud cloud-init on UC20 instead of restrict
On UC20, instead of restricting NoCloud to disable importing from arbitrary
filesystems, we want to just wholesale disable cloud-init for NoCloud after it
has run for first boot. The only situation we don't want to do that right now is
if there is a cloud.conf in the gadget snap, which indicates it is a "real
cloud", and should not be disabled. Actually enabling the cloud.conf case is a
followup commit, but doing this now should not appreciably expand cloud-init's
usage. We check the gadget's mount dir for existence of the cloud.conf because
the mount dir should be trusted, though if we wanted more assurance we could
manually look inside the snap's squashfs instead.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>