snap: auto-import will not try to auto-create users on managed devices
The `snap auto-import` code right will always try to create all
known system-users when it imports any assertions. However this
leads to systemd errors and a degraded boot when a device is already
managed and a removable device with a user assertion is attached
to the device.

This commit changes the auto-import code to only try to create
known users if the device is unmanaged. It will still import
assertions thought.

This fixes https://bugs.launchpad.net/newparis/+bug/1893331