tests/core/uc20-recovery: fix check for at least specific calls to mock-shutdown
The lines in the script currently actually look like:

```
+ cat /tmp/mock-shutdown.calls
-r +10 reboot scheduled to update the system
stop serial-getty@ttyS0
stop getty@tty1
-r +0 reboot scheduled to update the system
```

because on uc20 /usr/sbin/shutdown is a symlink to /bin/systemctl, so when we 
mount on top of it, we also end up mounting on top of /bin/systemctl because we
are mounting on top of a symlink. 

Thus, we only save invocations of mock-shutdown that start with "-r" to indicate
that a reboot was requested, and we then check for specifically the two reboots
that we expect to have been called.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>