o/devicestate/devicestate_cloudinit_test.go: fix various test issues
The tests were not checking the logged message, so the fact that the res.Action
returned from these tests were not the right ones ("disabled" instead of
"disable", etc.) was missed, so check the logged messages and update the
res.Action.

Also always setup a real gadget snap in snapstate with a gadget.yaml and
snap.yaml otherwise we get erroneous messages logged about missing those files
when getting the snap.Info for the gadget snap for the uc20 tests.

Finally, use DeepEquals on opts since there are multiple options in
CloudInitRestrictOptions now and we want to compare all of the options passed,
not just a single one now.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>