o/devicestate/devicemgr.go: always set DisableNoCloud = true for uc20
This is because even if a gadget snap is using NoCloud for it's cloud-init
config in cloud.conf, NoCloud is not very secure and not meant for production
systems, so we still disable it after the first boot anyways.

Also add a comment reflecting the slightly confusing nature of the
DisableNoCloud option.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>