tests: fix nested to work with qemu and kvm
The current code will clean the PARAM_MACHINE if running with qemu.
This is incorrect for two reasons:
1. nested kvm is possible with the right module parameter
2. it does not reset PARAM_CPU which leads to "-cpu host" but that
   will fail to start without using the accel=kvm option

This commit fixes both issues by checking if nested kvm is available
and uses it if it is. If not it resets both parameters. To test
run:
```
$ spread -v qemu-nested:ubuntu-20.04-64:tests/nested/manual/core-early-config
```
which did not work before this commit.