nested: add support to telnet to serial port in nested VM
This commit adds support to use telnet to access the nested VM
serial port. This is useful for interactive debugging of the
nested VM if e.g. the creation of a user fails. It is also useful
to test/play-with console-conf in the nested env.

With that inside the L1 qemu one can type:
```
$ telnet localhost 7777
```
to get access to the serial port.

With "ctrl-]" and "send escape\nq\n" the telnet session can be
stopped again.

As a drive-by it cleans a duplicated PARAM_SERIAL definition. We
should address the duplication of all the PARAM_* for qemu in a
followup.