tests/lib/nested.sh: if NESTED_CORE_CHANNEL is empty, don't specify --channel
For tests that will use non-dangerous uc20 models, we cannot override the
channel used, as the channel used has to come from the model.

To accommodate this, tests that will use non-dangerous models will need to set
NESTED_CORE_CHANNEL to empty. It remains to be seen how devices with such a
model would use tests that also currently use NESTED_CORE_CHANNEL, but we should
deal with that when necessary to unblock testing non-dangerous uc20 models here.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>