tests/lib/nested.sh: set UBUNTU_IMAGE_SNAP_CMD when building an image
This allows us to test nested VM's built with the most recent version of
"snap prepare-image" available, specifically the one from the spread run itself.

It also has the useful side effect of allowing us to build images with snap
declarations signed and provided by the fakestore, because this snap command
will have the trustkeys built into it.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>