tests/lib/store.sh: add make_snap_installable_with_id
This allows providing a .snap file and a specific snap ID, and creating
appropriate snap-revisions and snap-declarations in the provided store top-dir.

This is useful because oftentimes we want to use a specific snap ID with the
generated assertions, for example the snapd snap needs to have a specific snap
ID when building a UC20 image.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>