tests/nested/manual: add test for grades above signed booting with testkeys
This test exercises both the fakestore utilities and generated assertions, but
also that we can boot successfully with model assertions other than dangerous as
all our other tests use grade dangerous.

Because this test needs grades above dangerous, we cannot use unasserted snaps,
and need to use the fakestore to generate various snap-declaration and
snap-revision assertions during the course of assembling the image.

Using fakedevicesvc to get a serial assertion is currently disabled, as snapd is
unable to send a request for generating a serial assertion due to:

2020-09-03T21:35:05Z ERROR cannot deliver device serial request: invalid assertion: could not validate model assertion (no matching public key "EAD4DbLxK_kn0gzNCXOs3kd6DeMU3f-L6BEsSEuJGBqCORR0gXkdDxMbOm11mRFu" for signature by "developer1")

That will be fixed/enabled in a followup PR.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>