tests/lib/nested: bump the retry count
We have observed it takes very long for a nested VM with TPM and full disk
encryption, to be set up to the point that the user1 added by the test is picked
up by pam_extrausers. Only once that happens, logging over ssh is possible.

Investigating further, with the current retry count, the attemps are exhausted
before logging in over ssh is possible. Bump the retry count so that we get more
retries until we have better ideas on how to fix this.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>