configcore: do not error in console-conf.disable for install mode
During install mode the gadget-defaults will also be set as part
of the system install change. However during install mode
console-conf has no "complete" file, it just never runs
in install mode. So we need to detect this and do nothing
or the install mode will fail.

Note that this is not a great solution, ideally we would compare
what is set with the system default. But that requires some deeper
changes.