sysconfig,o/devicestate: mv DisableNoCloud to DisableAfterLocalDatasourcesRun
Also include "None" datasource in the definition of a local datasource on UC20
so it is disabled after first boot even when used.

We don't change the behavior on UC16/UC18 however, so on those platforms the
"None" datasource will just be restricted to only ever allow cloud-init to use
"None" as a datasource, so users of the "None" datasource with cloud-init on
Ubuntu Core will still be protected from the specific NoCloud attack from
CVE-2020-11933.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>