tests/lib/nested.sh: add sync's for shutdown helper to prevent file corruption
Sometimes we see that a file is written for a running VM, but upon shutting it
down and bringing it back up, that same file is empty. Notably this seems to
happen for the sudoers files we use when creating new users, so hopefully these
sync commands help protect against this happening.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>