tests: do not use su -l
The gdbserver test does not really need to use su -l for what is being
observed. The -l switch can cause a session to be spawned/removed
briefly and we try to avoid that, as rapid startup/shutdown is performed
asynchronously, and may interfere with other tests if test ordering and
timing are not in our favour.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>