tests/nested/manual/refresh-revert-fundamentals: disable temporarily
These tests currently always fail because the snapd snap in beta and newer
writes a modeenv which the initramfs in all published kernel snaps will corrupt
essentially, resulting in broken installs. To fix the issue we need a new kernel
snap, so until that happens mark the test as manual to allow landing PR's.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>