snap: add new `snap debug show-recovery-key` command
This command shows the recovery key in a format that is suitable
for the secboot.ParseRecoveryKey() helper that is used to unlock
the disk in a desaster recovery scenario.

This command is intended to be used by console-conf or similar
tools to give the user a chance to write down the recovery key.