tests/nested/manual/cloud-init-*: refactor to not re-create nested systemd unit
This is better because it doesn't involve recreating all the nested params, as
now all the qemu parameters (namely -cdrom) are the same across the reboot, just
that the actual backing file for the cdrom has changed to reflect the new image
there.

Additionally, this lets us use the more robust method of detecting reboots with
nested_wait_for_reboot which tracks the bootid.

Also use the new helper from nested.sh to build the seeds, for the never-used
variant of the test, the first cdrom image is essentially empty and ignored by
everything on the system, but is a placeholder for the second boot.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>