cmd/s-b/initramfs-mounts: use ConfigureTargetSystem for install, recover modes
For install mode, this will effectively do the same thing as just calling
DisableCloudInit directly, but for Recover mode since we provide the gadget snap
from the seed we opened, we will configure various things about the ephemeral
recover system using the gadget defaults such as disabling console-conf or SSH,
etc. in addition to disabling cloud-init.

Fixes: https://bugs.launchpad.net/snapd/+bug/1895856

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>