boot: fix debug bootloader variables dump on UC20 systems
The output of `snap debug boot-vars` is not that useful on UC20 systems.
Specifically, on Raspberry PI, the alternative location of bootloader
environment (boot.sel) is not even selected if none of the bootloader flags is
set, so running the debug command on such system outright fails.

Try to remedy this by picking a reasonable set of options based on either root
dir, of the flags.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>