update-pot: ignore .go files inside .git when running xgettext-go
If someone has a branch name that ends with ".go", then the file for that branch
in the .git folder will also be picked up, but it will not be a real .go file
and as such the processing in xgettext-go will fail like so:

2020/09/22 16:14:08 processFiles failed with: /root/parts/snapd-deb/build/.git/logs/refs/remotes/bboozzoo/bboozzoo/uc20-tweak-sid-packaging-params.go:1:1: expected 'package', found 'INT' 0000000000000000000000000000000000000000 (and 1 more errors)

With the git remote branch on github.com/bboozzoo/snapd,
"bboozzoo/uc20-tweak-sid-packaging-params.go".

This ensures that all files from .git are ignored when we process the i18n as
part of the build process.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>