Split the nested helper in 2 parts
The idea of this change is to split the nested helper which currently y
really big in 2 parts.

The first part is the nested-tool.sh which in a following pr will be
converted to a tool. This tools contains the preparation/restoration of
the environment and the vms, and some functions needed to support those.

The second part is the nested,sh which contains the common functions
used by the tests.

This is the first step of the migration to the nested tool.